backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
email-validator
python-jose[cryptography]
python-multipart
sqlalchemy
python-dotenv
argon2-cffi
